thread bantime deeper to use later

+++ src/Detection.zig
@@ -1,4 +1,4 @@
 hit: []const u8,
 heat: u16 = 1,
 decay: u16 = 0,
-ban_time: ?usize = null,
+ban_time: ?u32 = null,
